# backend/internal/models/todo_item.go
package models

import "errors"

type TodoItem struct {
	Id          int    `json:"id" db:"id"`
	Title       string `json:"title" db:"title" binding:"required"`
	Description string `json:"description" db:"description"`
	Status      string `json:"status" db:"status"`
}

type UpdateItemInput struct {
	Title       *string `json:"title"`
	Description *string `json:"description"`
	Status      *string `json:"status" db:"status"`
}

func (i UpdateItemInput) Validate() error {
	if i.Title == nil && i.Description == nil && i.Status == nil {
		return errors.New("update structure has no values")
	}

	return nil
}